% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress.data}
\alias{compress.data}
\title{compress.data}
\usage{
compress.data(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a list with the compressed profile x and associated repeat vector vrep
}
\description{
compress data and return a weighted profile
}
