% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpopw.R
\docType{package}
\name{fpopw}
\alias{fpopw}
\title{fpopw: A package to solve the optimal partionning and segment neighborhood problems using
a weighted L2-loss.}
\description{
The fpopw package provides wrapper to four functionnal pruning functions to
to solve the optimal partionning and segment neighborhood problems with the L2-loss:
Fpop, Fpop_w, Fpsn, Fpsn_w
}
\section{fpopw functions}{

fpopw functions are Fpop, Fpop_w, Fpsn, Fpsn_w, Fpsn_w_nomemory
}

