% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{get.change}
\alias{get.change}
\title{get.change}
\usage{
get.change(smt)
}
\arguments{
\item{smt}{smoothed profile}
}
\value{
a vector of changes including n
}
\description{
Function returning changes in a smoothed profile
}
