% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSelection.R
\name{getSegSums_}
\alias{getSegSums_}
\title{getSegSums_}
\usage{
getSegSums_(x, tau)
}
\arguments{
\item{x}{data}

\item{tau}{changes (including n)}
}
\value{
a vector of the sums
}
\description{
A function to get the segment sums of a vector given some changes including n
}
