% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSelection.R
\name{getSMT}
\alias{getSMT}
\title{getSMT}
\usage{
getSMT(res, K = NULL)
}
\arguments{
\item{res}{output of Fpop, Fpop_w, Fpsn or Fpsn_w}

\item{K}{the number of changes (only if Fpsn or Fpsn_w)}
}
\value{
a vector of the smoothed profile
}
\description{
A function to get the smoothed profile from the output of Fpop, Fpop_w, Fpsn and Fpsn_w
}
