% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSelection.R
\name{getSMT_}
\alias{getSMT_}
\title{getSMT_}
\usage{
getSMT_(x, weights = NULL, tauHat)
}
\arguments{
\item{x}{data}

\item{weights}{weights}

\item{tauHat}{changes (including n)}
}
\value{
a vector of the smoothed profile
}
\description{
A function to get the smoothed profile from the data, weights and changepoints
}
