% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSelection.R
\name{getTau_nomemory}
\alias{getTau_nomemory}
\title{getTau_nomemory}
\usage{
getTau_nomemory(res_fpsn, K_selected)
}
\arguments{
\item{res_fpsn}{output of the function res_fpsn_nomemory}

\item{K_selected}{K obtained using select_Fpsn}
}
\value{
return a set of changes
}
\description{
function to recover changes for a given selected K after fpsn_nomemory
}
\examples{
x <- c(rnorm(100), rnorm(10^3)+2, rnorm(1000)+1)
res <- Fpsn_w_nomemory(x=x, w=rep(1, length(x)), K=100)
select.res <- select_Fpsn(res, method="givenVariance")
tau <- getTau_nomemory(res, select.res)
smt <- getSMT_(res$signal, res$weights, tau)
}
