% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpaccess.R
\name{retour_op}
\alias{retour_op}
\title{retour_op}
\usage{
retour_op(path, i)
}
\arguments{
\item{path}{vector of length n containing the best last changes for any j in \eqn{[1, n]}. This vector is computed in the Fpop and Fpop_w using the colibri_op_c or colibri_op_weight_c function.}

\item{i}{the last position to consider to start the backtracking.}
}
\value{
set of optimal changes up to i.
}
\description{
Function used internally by Fpop and Fpop_w to do the backtracking and recover the best set of changes from 1 to i
}
