% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpaccess.R
\name{retour_sn}
\alias{retour_sn}
\title{retour_sn}
\usage{
retour_sn(path)
}
\arguments{
\item{path}{matrix of size (K x n) containing the last optimal changes up to j in k segments with i in \eqn{[1, n]} and k in \eqn{[1, K]}. This matrix is computed in the Fpsn or Fpsn_w function using the colibri_sn_c or colibri_sn_weight_c functions.}
}
\value{
a matrix of size (K x K) containing the best segmentations in 1 to K segments.
}
\description{
Function used internally by Fpsn and Fpsn_w to do the backtracking and recover the best set of segmentations in 1 to K changes from 1 to n.
}
