% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSelection.R
\name{saut}
\alias{saut}
\title{saut}
\usage{
saut(Lv, pen, Kseq, n, seuil = sqrt(n)/log(n), biggest = TRUE)
}
\arguments{
\item{Lv}{likelihood}

\item{pen}{penalty}

\item{Kseq}{number of changes}

\item{n}{number of datapoints}

\item{seuil}{threshold}

\item{biggest}{heuristic (biggest jump or slope)}
}
\value{
a selected number of chagnes
}
\description{
model selection function taken from S3IB,
}
