% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSelection.R
\name{sdDiff}
\alias{sdDiff}
\title{sdDiff}
\usage{
sdDiff(x, method = "MAD")
}
\arguments{
\item{x}{signal}

\item{method}{used to estimate the variance : MAD or HALL}
}
\value{
return a numeric value
}
\description{
Function to estimate the standard deviation
}
