% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{uncompress.vec}
\alias{uncompress.vec}
\title{uncompress.vec}
\usage{
uncompress.vec(vec.rep)
}
\arguments{
\item{vec.rep}{integer vector with the number of time each point should be repeated}
}
\value{
return a vector to uncompress a profile, segmentation or smt
}
\description{
return a vector to uncompress a profile, segmentation or smt
}
